driver: gpio: Unified naming for k1 chip
community inclusion
category: feature
bugzilla: https://github.com/RVCK-Project/rvck/issues/161

--------------------------------

In the previously submitted code, there was an issue
with the inconsistent use of the names 'k1' and 'k1x'.
This has now been standardized to 'k1'.
