+++ drivers/gpio/Kconfig
@@ -1818,12 +1818,12 @@ config GPIO_SIM
 	  This enables the GPIO simulator - a configfs-based GPIO testing
 	  driver.
 
-config GPIO_K1X
-	bool "Spacemit k1x GPIO support"
-	depends on PINCTRL_SPACEMIT_K1X
+config GPIO_K1
+	bool "Spacemit k1 GPIO support"
+	depends on PINCTRL_SPACEMIT_K1
 	help
-	  Say yes here to support the k1x GPIO device.
-	  The k1x GPIO device may have several banks, and each
+	  Say yes here to support the k1 GPIO device.
+	  The k1 GPIO device may have several banks, and each
 	  bank control at most 32 GPIO pins. The number of banks
 	  is passed by device tree or platform data.
